refactor: Generalização
