+++ src/main/java/com/logistic/services/generic/IGenericDeleteService.java
@@ -1,0 +1,10 @@
+package com.logistic.services.generic;
+
+import org.springframework.transaction.annotation.Transactional;
+	
+public interface IGenericDeleteService<T> extends IGenericUpdateService<T> {
+	
+	@Transactional
+	void delete(Integer id);
+
+}
