+++ src/main/java/com/logistic/services/generic/IGenericUpdateService.java
@@ -1,0 +1,10 @@
+package com.logistic.services.generic;
+
+import org.springframework.transaction.annotation.Transactional;
+	
+public interface IGenericUpdateService<T> extends IGenericInsertService<T> {
+	
+	@Transactional
+	T update(T obj);
+
+}
