+++ src/main/java/com/logistic/services/imp/PedidoService.java
@@ -4,13 +4,11 @@ import java.util.List;
 import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.stereotype.Service;
 
 import com.logistic.domain.Pedido;
 import com.logistic.repositories.PedidoRepository;
 import com.logistic.services.IPedidoService;
-import com.logistic.services.exception.DataIntegrityException;
 import com.logistic.services.exception.ObjectNotFoundException;
 
 @Service
@@ -23,27 +21,6 @@ public class PedidoService implements IPedidoService {
 	public Pedido find(Integer id) {
 		Optional<Pedido> obj = repo.findById(id);
 		return obj.orElseThrow(() -> new ObjectNotFoundException("Objeto não encontrado! Id: " + id + ", Tipo: " + Pedido.class.getName()));
-	}
-	
-	@Override
-	public Pedido insert(Pedido obj) {
-		obj.setId(null);
-		return repo.save(obj);
-	}
-	
-	@Override
-	public Pedido update(Pedido obj) {
-		return repo.save(find(obj.getId()));
-	}
-
-	@Override
-	public void delete(Integer id) {
-		find(id);
-		try {
-			repo.deleteById(id);
-		} catch (DataIntegrityViolationException e) {
-			throw new DataIntegrityException("Não é possivel excluir uma categoria que possui produtos");
-		}
 	}
 	
 	@Override
